requirements.txt
# Python Backend Dependencies for redSpec.AI

# Google ADK and AI
google-adk>=1.18.0
google-genai
google-cloud-aiplatform

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart

# Environment & Config
python-dotenv>=1.0.0

# Async & WebSockets
websockets>=12.0
aiofiles>=23.0.0

# Git Integration
GitPython>=3.1.40

# Data & JSON
pydantic>=2.0.0
pydantic-settings

# CORS
fastapi-cors

# SSE (Server-Sent Events)
sse-starlette>=1.8.0
